refactor: change learn lesson page structure
