+++ front/src/components/lessonBlocks/OngoingShort/OngoingShort.desktop.jsx
@@ -50,7 +50,6 @@ OngoingShortDesktop.propTypes = {
   lesson: PropTypes.shape({
     id: PropTypes.number.isRequired,
     name: PropTypes.string.isRequired,
-    maintainer: PropTypes.string.isRequired,
     description: PropTypes.string.isRequired,
     percentage: PropTypes.number.isRequired,
   }).isRequired,
