+++ front/src/components/lessonBlocks/OngoingShort/OngoingShort.jsx
@@ -20,7 +20,6 @@ OngoingShort.propTypes = {
   lesson: PropTypes.shape({
     id: PropTypes.number.isRequired,
     name: PropTypes.string.isRequired,
-    maintainer: PropTypes.string.isRequired,
     description: PropTypes.string.isRequired,
   }).isRequired,
 };
