+++ front/src/pages/User/Home/OngoingLessons/constants.js
@@ -1,1 +1,1 @@
-export const PAGE_SHORT_SIZE = 10;
+export const PAGE_SHORT_SIZE = 3;
