--- front/src/pages/User/LessonPage/Block/index.js
@@ -1,3 +1,0 @@
-import Block from './Block';
-
-export default Block;
