+++ front/src/pages/User/LessonPage/BlockElement/QuizBlockResult/Answer/index.js
@@ -1,0 +1,3 @@
+import Answer from './Answer';
+
+export default Answer;
