+++ front/src/pages/User/LessonPage/BlockElement/QuizBlockResult/AnswerResult/index.js
@@ -1,0 +1,3 @@
+import AnswerResult from './AnswerResult';
+
+export default AnswerResult;
