+++ front/src/pages/User/LessonPage/BlockElement/QuizBlockResult/Question/Question.jsx
@@ -1,0 +1,16 @@
+import PropTypes from 'prop-types';
+import { Col, Typography } from 'antd';
+
+const { Text } = Typography;
+
+const Question = ({ text }) => (
+  <Col span={24}>
+    <Text style={{ fontStyle: 'italic' }}>{text}</Text>
+  </Col>
+);
+
+Question.propTypes = {
+  text: PropTypes.string.isRequired,
+};
+
+export default Question;
