+++ front/src/pages/User/LessonPage/BlockElement/QuizBlockResult/Question/index.js
@@ -1,0 +1,3 @@
+import Question from './Question';
+
+export default Question;
