+++ front/src/pages/User/LessonPage/BlockElement/QuizBlockResult/QuizblockResult.styled.js
