+++ front/src/pages/User/LessonPage/BlockElement/index.js
@@ -1,0 +1,3 @@
+import BlockElement from './BlockElement';
+
+export default BlockElement;
