+++ front/src/pages/User/LessonPage/GroupBlock/GroupBlock.styled.js
@@ -1,6 +1,16 @@
 import styled from 'styled-components';
 import { Col, Row } from 'antd';
 
+export const BlockCol = styled(Col)`
+  display: flex;
+  justify-content: center;
+  padding-top: ${(props) => (props.top ? props.top : '3rem')};
+`;
+
+export const PageRow = styled(Row)`
+  width: 100%;
+`;
+
 export const BlockWrapper = styled(Row)`
   width: 100%;
   max-width: 614px;
@@ -9,16 +19,6 @@ export const BlockWrapper = styled(Row)`
   border-radius: 8px;
 `;
 
-export const PageRow = styled(Row)`
-  width: 100%;
-`;
-
 export const StyledRow = styled(Row)`
   width: 100%;
-`;
-
-export const BlockCol = styled(Col)`
-  display: flex;
-  justify-content: center;
-  padding-top: ${(props) => (props.top ? props.top : '3rem')};
 `;
