+++ front/src/pages/User/LessonPage/LessonPage.jsx
@@ -1,176 +1,65 @@
-import { useEffect, useState } from 'react';
-
-import { useHistory, useParams } from 'react-router-dom';
-import { Col, Progress, Row, Typography } from 'antd';
+import { Col, Row } from 'antd';
 import { useTranslation } from 'react-i18next';
-import { useMutation, useQuery } from 'react-query';
-import { LESSON_BASE_QUERY } from '@sb-ui/utils/queries';
-import {
-  generateBlockByElement,
-  groupBlocks,
-  prepareResultToAnswers,
-} from '@sb-ui/pages/User/LessonPage/utils';
-import Block from '@sb-ui/pages/User/LessonPage/Block';
-import { getLessonById, postLessonById } from '@sb-ui/utils/api/v1/student';
-import { USER_HOME } from '@sb-ui/utils/paths';
-import QuizBlock from '@sb-ui/pages/User/LessonPage/QuizBlock';
-import InfoBlock from './InfoBlock';
-import * as S from './LessonPage.styled';
 
-const { Text } = Typography;
+import QuizBlock from '@sb-ui/pages/User/LessonPage/QuizBlock';
+import Header from '@sb-ui/pages/User/LessonPage/Header';
+
+import GroupBlock from '@sb-ui/pages/User/LessonPage/GroupBlock';
+import BlockElement from '@sb-ui/pages/User/LessonPage/BlockElement';
+
+import { useCallback } from 'react';
+import * as S from './LessonPage.styled';
+import { useLesson } from './useLesson';
 
 const LessonPage = () => {
   const { t } = useTranslation();
-  const history = useHistory();
+  const {
+    lesson,
+    blocks,
+    total,
+    nextCount,
+    isLoading,
+    isFinal,
+    interactiveBlock,
+    setQuizAnswer,
+    handleSendClick,
+    handleFinishClick,
+    handleNextClick,
+    handleStartClick,
+  } = useLesson();
 
-  const { id: lessonId } = useParams();
-  const [blocks, setBlocks] = useState([]);
-  const [interactiveBlock, setInteractiveBlock] = useState();
-  const [quizAnswer, setQuizAnswer] = useState({});
-  const [isFinal, setIsFinal] = useState(false);
-  const [total, setTotal] = useState(0);
-  const [nextCount, setNextCount] = useState(0);
-
-  const { data: responseData, isLoading } = useQuery(
-    [
-      LESSON_BASE_QUERY,
-      {
-        id: lessonId,
-      },
-    ],
-    getLessonById,
+  const isQuizBlockResult = useCallback(
+    (block) =>
+      block.type === 'quiz' &&
+      block.content.data?.answers?.every((x) => x.correct === undefined),
+    [],
   );
-
-  const { mutate } = useMutation(postLessonById, {
-    onSuccess: (newBlocks) => {
-      const { isFinal: finishStatus, lesson, total: totalBlocks } = newBlocks;
-      if (finishStatus) {
-        setIsFinal(finishStatus);
-      }
-      const answers = lesson?.userAnswer?.response;
-
-      const dataBlocks = lesson?.blocks;
-      if (dataBlocks) {
-        if (lesson?.blocks?.length === totalBlocks) {
-          setNextCount(dataBlocks.filter((x) => x.type === 'next').length);
-          setBlocks([...groupBlocks(dataBlocks)]);
-          setInteractiveBlock(null);
-        } else {
-          setNextCount(
-            (prev) => prev + dataBlocks.filter((x) => x.type === 'next').length,
-          );
-          setBlocks((prev) => {
-            const newPrev = JSON.parse(JSON.stringify(prev));
-            if (answers) {
-              newPrev[newPrev.length - 1][0].content.data.answers = newPrev[
-                newPrev.length - 1
-              ]?.[0]?.content?.data?.answers?.map((x, i) => ({
-                ...x,
-                correct: answers[i],
-              }));
-              newPrev[newPrev.length - 1][0].answer = {
-                results: lesson?.answer?.results,
-              };
-            }
-            return [...newPrev, ...groupBlocks(dataBlocks)];
-          });
-          const lastBlock = dataBlocks[dataBlocks.length - 1];
-          if (lastBlock) {
-            setInteractiveBlock(lastBlock);
-          }
-        }
-      }
-    },
-  });
-
-  useEffect(() => {
-    const preparedData = prepareResultToAnswers(responseData);
-    const dataBlocks = preparedData?.lesson?.blocks;
-    const finishStatus = preparedData?.isFinal;
-    if (finishStatus) {
-      setIsFinal(finishStatus);
-    }
-    const totalData = preparedData?.total;
-    setTotal(totalData);
-
-    if (dataBlocks) {
-      setBlocks(groupBlocks(dataBlocks));
-      setNextCount(dataBlocks.filter((x) => x.type === 'next').length);
-      if (totalData === dataBlocks.length) {
-        setInteractiveBlock(null);
-      } else {
-        const lastBlock = dataBlocks[dataBlocks.length - 1];
-        if (lastBlock) {
-          setInteractiveBlock(lastBlock);
-        }
-      }
-    }
-  }, [responseData]);
-
-  const { lesson } = responseData || {};
-
-  const handleStartClick = () => {
-    mutate({ id: lessonId, action: 'start' });
-  };
-
-  const handleNextClick = () => {
-    const { revision, blockId } = interactiveBlock;
-    mutate({ id: lessonId, action: 'next', revision, blockId });
-  };
-
-  const handleFinishClick = () => {
-    mutate({ id: lessonId, action: 'finish' });
-    history.push(USER_HOME);
-  };
-
-  const handleSendClick = () => {
-    const { revision, blockId } = interactiveBlock;
-    mutate({
-      id: lessonId,
-      action: 'response',
-      revision,
-      blockId,
-      data: {
-        response: interactiveBlock?.content?.data?.answers.map(
-          (x, i) => !!quizAnswer.includes(i),
-        ),
-      },
-    });
-  };
 
   return (
     <S.Page>
       <S.GlobalStylesLessonPage />
-      <S.ProgressCol span={24}>
-        <Progress
-          showInfo={false}
-          percent={(blocks?.flat()?.length / (total - nextCount)) * 100}
-          strokeWidth={2}
-          strokeLinecap="round"
-        />
-      </S.ProgressCol>
-      <S.PageRow justify="center" align="top">
-        <S.BlockCol
-          xs={{ span: 20 }}
-          sm={{ span: 18 }}
-          md={{ span: 16 }}
-          lg={{ span: 14 }}
-        >
-          <InfoBlock
-            isLoading={isLoading}
-            total={total - nextCount}
-            lesson={lesson}
-          />
-        </S.BlockCol>
-      </S.PageRow>
+      <Header
+        lesson={lesson}
+        percent={(blocks?.flat()?.length / (total - nextCount)) * 100}
+        total={total - nextCount}
+        isLoading={isLoading}
+      />
 
-      {blocks?.map((block) => (
-        <Block>
-          {block
-            .map((element) => generateBlockByElement(element))
+      {blocks.map((groupBlock) => (
+        <GroupBlock
+          key={groupBlock?.map((x) => x.blockId).join('')}
+          elements={groupBlock
+            .map((block) => {
+              if (isQuizBlockResult(block)) {
+                return null;
+              }
+
+              return <BlockElement element={block} />;
+            })
             .filter((x) => !!x)}
-        </Block>
+        />
       ))}
+
       {!isLoading &&
         (isFinal ||
           (interactiveBlock?.type === 'next' && !isFinal) ||
@@ -211,19 +100,20 @@ const LessonPage = () => {
               md={{ span: 16 }}
               lg={{ span: 14 }}
             >
-              <Block isQuiz>
-                <Text
-                  style={{
-                    fontStyle: 'italic',
-                  }}
-                >
+              <GroupBlock>
+                <S.TextItalic>
                   {interactiveBlock?.content?.data?.question}
-                </Text>
-              </Block>
+                </S.TextItalic>
+              </GroupBlock>
             </Col>
           </Row>
           <S.RowQuiz justify="center" align="top">
-            <Col span={24}>
+            <Col
+              xs={{ span: 24 }}
+              sm={{ span: 24 }}
+              md={{ span: 16 }}
+              lg={{ span: 14 }}
+            >
               <QuizBlock
                 key={interactiveBlock?.blockId}
                 setQuiz={setQuizAnswer}
@@ -237,7 +127,7 @@ const LessonPage = () => {
               lg={{ span: 14 }}
             >
               <S.LessonButtonSend onClick={handleSendClick}>
-                Send
+                {t('lesson.send')}
               </S.LessonButtonSend>
             </S.SendWrapper>
           </S.RowQuiz>
