+++ front/src/pages/User/LessonPage/QuizBlock/Quizblock.styled.js
@@ -1,17 +1,8 @@
 import styled from 'styled-components';
 import { Checkbox, Col, Row } from 'antd';
 
-export const BlockWrapper = styled(Row)`
-  width: 100%;
-  max-width: 614px;
-  padding: 2rem;
-  background-color: rgba(245, 245, 245, 1);
-  border-radius: 8px;
-`;
-
 export const BlockWrapperWhite = styled(Row)`
   width: 100%;
-
   padding: 2rem;
   background-color: white;
   box-shadow: 0px -4px 10px rgba(231, 231, 231, 0.5);
